// lib/services/front-types.d.ts
/*
 Copyright 2016-2017 Resin.io

 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at

 http://www.apache.org/licenses/LICENSE-2.0

 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */

import { Front } from 'front-sdk';
import { ServiceAPIHandle, ServiceEmitContext } from './service-types';

export interface FrontCommentEmitContext extends ServiceEmitContext {
    conversation_id: string;
    author_id: string;
    body: string;
    options: {
        archive: false;
    };
    sender: {
        handle: string;
    };
    subject: string;
    type: 'message' | 'comment';
}
export interface FrontConversationEmitContext extends ServiceEmitContext {
    author_id: string;
    body: string;
    channel_id: string;
    metadata: {
        thread_ref: string;
    };
    options: {
        archive: false;
    };
    sender: {
        handle: string;
    };
    subject: string;
    to: string[];
    type: 'conversation';
}
export interface FrontHandle extends ServiceAPIHandle {
    front: Front;
}